fix(desktop): boot after ready to break Electron ESM startup deadlock
Electron ESM emits `ready` only after the main module finishes
evaluating, so a top-level `await app.whenReady()` in the startup chain
deadlocks: module evaluation waits on ready, ready waits on module
evaluation. The storage-root repair dialog hit exactly this — any
root-identity conflict hung the process silently with no window.

Split the entry: main.ts now does only pre-ready work (setName, E2E
userData redirect, single-instance lock with a proper return) and
dynamic-imports boot.ts inside the whenReady callback. boot.ts keeps the
whole startup chain (root-identity check, stores, IPC, lifecycle) as
module-level code after ready, so the check still precedes every store
and db write and confirmRepair no longer needs whenReady at all.

Also:
- fix losing-second-instance exiting without returning, so it never
  touches shared state (was opening SQLite before exit)
- surface fatal boot errors via showErrorBox instead of a silent exit
- E2E regression test: conflicting storage root reaches ready with the
  repair dialog open and writes nothing before the user answers
